+++ R/mixvlmc-package.R
@@ -23,6 +23,8 @@
 #' @importFrom Rcpp evalCpp
 #' @importFrom stats simulate
 #' @importFrom stats predict
+#' @importFrom ggplot2 autoplot
+#' @importFrom rlang .data
 #' @useDynLib mixvlmc
 ## usethis namespace: end
 NULL